完善AI基础配置，完善Service接口

+++ src/main/java/org/cjj/saaassistant/pojo/Product.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -12,7 +13,7 @@ import java.util.Date;
 public class Product {
     private int productId;
     private String productName;
-    private Integer productPrice;
+    private BigDecimal productPrice;
     private Integer productStock; // 库存数量
     private Date createTime; // 创建时间
     private Date updateTime; // 更新时间
